requirements.txt
PyYAML~=6.0.2
opencv-python~=4.10.0.84
numpy~=2.2.6
pillow~=11.2.1
sahi~=0.11.23
torch~=2.7.0+cu128
efficientnet_pytorch~=0.7.1
pytest~=8.3.5
efficientnet_pytorch~=0.7.1